Added string conbiner

+++ pca10056/blank/ses/src/qspi_flash.h
@@ -4,7 +4,7 @@
 
 extern bool  qspi_test(void);
 extern void  read_REMS(uint8_t*, uint8_t*);
-extern bool  qspi_read_id(uint16_t *manufacturer_id_readback);
+extern bool  qspi_read_id(uint8_t *manufacturer_id_readback);
 
 
 #endif
